requirements.txt
requests>=2.18.4
e-objects>=1.2.0
beautifulsoup4>=4.8.0